// Desktop/cs5500-gui-arch-huangweili97-main/src/types/types.tsx
export interface QuestionProps {
    qid: string;
    title: string;
    text: string;
    tagIds: string[];
    askedBy: string;
    askDate: Date;
    ansIds: string[];
    views: number;
    getAnswerCount: () => number;
    getQuestionViews: () => number;
    getTagsId: () => string[];
    calculateTimeElapsed: () => string;
}

/**
 * An interface representing the input properties of a component
 * displaying a list of questions based on a 
 * search string, a page index, and an order
 */
export interface QuestionPageProps {
    title_text?: string;
    qlist: QuestionProps[];
    qSize: number,
    search?: string;
    pageNum: number,
    getTagById: (id: string) => Tag | null;
    setQuestionOrder?: (order: string) => void;
    clickTag?: (tagName: string) => void;
    handleAnswer?: (qid: string) => void;
    handleNewQuestion?: () => void;
    setQuestionPage?: SetQuestionPage;
}
// ...
export type SetQuestionPage = (pageIndex: number, search: string, title: string) => void;